pyproject.toml
[project]
name = "objectmotionnotify"
version = "0.1.0"
description = "Object motion detection with open cv and notifier with telegram or email"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
]
